# openapi.yaml
openapi: "3.1.0"
info:
  title: Lumenqraph API
  version: "0.1.0"
  description: |
    Open-source, self-hostable event indexer for Soroban smart contracts on Stellar.
    Exposes decoded contract events, materialized token transfers, on-chain
    contract interfaces, read-only view-function calls, and webhook management
    over a plain REST API.
  license:
    name: MIT
    url: https://github.com/Lumen-Scribe/Lumenqraph/blob/main/LICENSE
  contact:
    url: https://github.com/Lumen-Scribe/Lumenqraph
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyHeader: []
  - ApiKeyBearer: []
  - {}

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
    ApiKeyBearer:
      type: http
      scheme: bearer

  # ---------------------------------------------------------------------------
  # Shared schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ---- pagination (#445) ----
    PageMeta:
      type: object
      required: [has_more, next_cursor]
      properties:
        has_more:
          type: boolean
          description: Whether more rows are available after this page.
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor to pass as `after` for the next page. Null on the last page.
...
    MaterializedPage:
      description: Page of materialized swap, NFT or liquidity rows.
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                type: object
                additionalProperties: true
...
  parameters:
    ContractId:
      name: contract_id
      in: path
      required: true
      schema:
        type: string
      description: Bech32m `C…` contract address.
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
      description: Maximum number of results to return.
    Offset:
      name: offset
      in: query
      deprecated: true
      schema:
        type: integer
        minimum: 0
        maximum: 10000
      description: |
        Deprecated: number of results to skip. Capped at 10000 (400 above
        that). Responses to offset requests carry `Deprecation: true` and
        `Warning` headers. Ignored when `after` is present. Use `after`.
    After:
      name: after
      in: query
      schema:
        type: string
      description: Opaque keyset cursor; the `next_cursor` of the previous page.

  # ---- reusable responses ----
  responses:
...
    BadRequest:
      description: Invalid parameter (e.g. malformed cursor or offset above the cap).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /contracts/{contract_id}/nfts:
    get:
      operationId: listContractNftEvents
      summary: Materialized NFT mint/transfer/burn events, newest first.
      tags: [transfers]
      parameters:
        - $ref: "#/components/parameters/ContractId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/After"
        - name: kind
          in: query
          schema:
            type: string
            enum: [mint, transfer, burn]
          description: Filter by event kind.
        - name: from
          in: query
          schema:
            type: string
          description: Filter by sender address.
        - name: to
          in: query
          schema:
            type: string
          description: Filter by recipient address.
        - name: token_id
          in: query
          schema:
            type: string
          description: Filter by token id.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterializedPage"
        "400":
          $ref: "#/components/responses/BadRequest"
  /contracts/{contract_id}/liquidity:
    get:
      operationId: listContractLiquidityEvents
      summary: Materialized liquidity add/remove events, newest first.
      tags: [transfers]
      parameters:
        - $ref: "#/components/parameters/ContractId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/After"
        - name: kind
          in: query
          schema:
            type: string
            enum: [add, remove]
          description: Filter by event kind.
        - name: provider
          in: query
          schema:
            type: string
          description: Filter by liquidity provider.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterializedPage"
        "400":
          $ref: "#/components/responses/BadRequest"
...